refactor: adjusting project architecture
