+++ src/api/schema/schema.ts
@@ -1,9 +1,9 @@
 import 'graphql-import-node';
 import { GraphQLSchema } from 'graphql';
 import { buildSchemaSync } from 'type-graphql';
-import { HelloResolver } from './modules/hello/hello.resolver';
-import { UserResolver } from './modules/user/user.resolver';
-import { LoginResolver } from './modules/login/login.resolver';
+import { HelloResolver } from './hello/hello.resolver';
+import { UserResolver } from './user/user.resolver';
+import { LoginResolver } from './login/login.resolver';
 
 // Load resolvers
 export const Schema: GraphQLSchema = buildSchemaSync({
